requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart>=0.0.9
openai>=1.0
anthropic>=0.25
twilio>=8.0
pypdf>=4.0
redis==5.2.1
rq==2.1.0
httpx==0.28.1
azure-storage-blob==12.24.0
faster-whisper==1.1.0
